12010 fix mt_srand error

+++ core/class/template.class.php
@@ -68,7 +68,7 @@ class template
         $limit = $attributes['limit']  ? count($letters) - 1 : 10;
 
         // Tirage de l'opération
-        mt_srand((float) microtime() * 1000000);
+        mt_srand((int) microtime() * 1000000);
         // Captcha simple limité à l'addition
         $operator = $attributes['limit'] ?  mt_rand(1, 4) : 1;
 
@@ -78,9 +78,9 @@ class template
         }
 
         // Tirage des nombres
-        mt_srand( microtime(true) * 1000000);
+        mt_srand( (int) microtime(true) * 1000000);
         $firstNumber = mt_rand(1, $limit);
-        mt_srand( microtime(true) * 1000000);
+        mt_srand( (int) microtime(true) * 1000000);
         $secondNumber = mt_rand(1, $limit);
 
         // Permutation si addition ou soustraction
@@ -110,7 +110,7 @@ class template
                 for ($i = 1; $i <= $firstNumber; $i++) {
                     $limit = $limit2[$i - 1];
                 }
-                mt_srand((float) microtime() * 1000000);
+                mt_srand((int) microtime() * 1000000);
                 $secondNumber = mt_rand(1, $limit);
                 $firstNumber =  $firstNumber * $secondNumber;
                 $result = $firstNumber / $secondNumber;
